Fix test after pyarrow-odbc fix.

+++ tests/test_table_hooks/test_pandas_hook.py
@@ -258,7 +258,7 @@ class TestPandasTableHookArrow:
                 "int32": pd.array(Values.INT32, dtype="int32[pyarrow]"),
                 "int64": pd.array(Values.INT64, dtype="int64[pyarrow]"),
                 "float32": pd.array(Values.FLOAT32, dtype="float[pyarrow]"),
-                "float64": pd.array(Values.FLOAT32, dtype="double[pyarrow]"),
+                "float64": pd.array(Values.FLOAT64, dtype="double[pyarrow]"),
                 "str": pd.array(Values.STR, dtype=pd.ArrowDtype(pa.string())),
                 "boolean": pd.array(Values.BOOLEAN, dtype="bool[pyarrow]"),
                 "date": pd.array(Values.DATE, dtype=pd.ArrowDtype(pa.date32())),
@@ -274,7 +274,7 @@ class TestPandasTableHookArrow:
         @materialize(input_type=(pd.DataFrame, "arrow"))
         def assert_expected(in_df):
             if ConfigContext.get().store.table_store.engine.dialect.name == "mssql":
-                # these are unavoidable differences when using bcp for writing tables
+                # # these are unavoidable differences when using bcp for writing tables
                 df["str"] = df["str"].replace("", pd.NA)
                 df["str"] = (
                     df["str"]
